+++ changelog.md
@@ -26,8 +26,10 @@
 
 # Unreleased
 
+- Added a check when displaying the item to prevent that if the slot is outside the slots this triggers an error but just a message in the console
 - Fix error with 1.20.4
 - Fix error with Meta
+- Fix error with slot page and else button
 
 # 1.0.1.9